feat: 优化代码

+++ entity/comment.go
@@ -17,8 +17,8 @@ type Comment struct {
 }
 
 type ArticlesWebCommentCount struct {
-	ArticleId    int `json:"article_id"`
-	CommentCount int `json:"comment_count"`
+	ArticleId    int   `json:"article_id"`
+	CommentCount int64 `json:"comment_count"`
 }
 
 func (*Comment) TableName() string {
